buffetiser/requirements.txt
django
django-cors-headers
django-redis
dj-database-url
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-rest-swagger
requests
beautifulsoup4
aiohttp
gunicorn
schedule
pytz
click